requirements.txt
Django>=4.0.1,<4.1.0
djangorestframework>=3.13.1,<3.14.0
psycopg2>=2.9.3,<2.10.0

tzdata>=2021.5,<2021.6